requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
bcrypt==4.0.1
python-multipart==0.0.6
gunicorn==21.2.0
aiohttp==3.11.1